+++ inits/drupal10.js
@@ -5,4 +5,9 @@
  */
 module.exports = {
   name: 'drupal10',
+  defaults: {
+    'php': '8.1',
+    'drush': '^11',
+    'composer_version': '2-latest',
+  },
 };