+++ inits/drupal7.js
@@ -5,4 +5,7 @@
  */
 module.exports = {
   name: 'drupal7',
+  defaults: {
+    'php': '7.4',
+  },
 };